pyproject.toml
 [build-system]
 requires = ["poetry-core>=1.9.0"]
 build-backend = "poetry.core.masonry.api"

 [tool.poetry]
 name = "codexctl"
 version = "0.1.1"
 description = "Manage containerized projects and per-run tasks with Podman"
 readme = "README.md"
 license = "Apache-2.0 OR MIT"
 authors = ["codexctl maintainers"]
 keywords = ["podman", "containers", "tasks", "developer-tools", "cli"]
 classifiers = [
   "Programming Language :: Python :: 3",
   "License :: OSI Approved :: MIT License",
   "Environment :: Console",
   "Operating System :: POSIX :: Linux",
   "Topic :: System :: Systems Administration",
 ]

 # Use the src/ layout
 packages = [
   { include = "codexctl", from = "src" },
 ]

 # Extra files to ship alongside the package (similar to the old MANIFEST.in
 # and setuptools data-files). These will be present in the sdist and wheel,
 # but note that Poetry does not support installing them under /usr/share;
 # they will live under the installed distribution instead.
 include = [
   { path = "completions/codexctl" },
   { path = "docs/*.md" },
   { path = "examples/**" },
 ]

 [tool.poetry.dependencies]
 python = ">=3.9,<4.0"
 PyYAML = ">=6.0"
 textual = ">=6.6"
 argcomplete = ">=3.1"
 
 [tool.poetry.group.dev.dependencies]
 pydevd-pycharm = "^253.28294.166"

[tool.poetry.scripts]
codexctl = "codexctl.cli.main:main"
codexctl-tui = "codexctl.tui.app:main"
